// src/data/api/generated/models/supplier.ts
/* eslint-disable */
// @ts-nocheck

export interface Supplier {
  /** @nullable */
  contact_name?: string | null;
  created_at: string;
  /** @nullable */
  email?: string | null;
  id: string;
  is_active: boolean;
  name: string;
  org_id: string;
  /** @nullable */
  phone?: string | null;
  updated_at: string;
}